Organize source files and update background
